🚀 working endpoint at http://localhost:8080/BooksApi-web/api/books
Crap this was harder than it needed to be. See following links for info.
https://docs.oracle.com/cd/E24329_01/web.1211/e24983/develop.htm#RESTF125
https://docs.oracle.com/cd/E24329_01/web.1211/e24983/configure.htm#RESTF182

I struggled hard to recognize, that "BooksApi-web" has to be in the url,
this is the context-root. You can find it in the application.xml in the
ear module. No idea yet how to change it. The BooksApiApplication
defines the "api" after the context root.